// frontend/features/api/model/organizationResource.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * Petrol Integrity System API Documentation
 * OpenAPI spec version: 0.0.1
 */

export interface OrganizationResource {
  id: string;
  name: string;
  slug: string;
  status: string;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
}